requirements.txt
aiohttp==3.9.5
APScheduler==3.10.4
Flask==3.0.3
Flask_JWT_Extended==4.6.0
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
pytz==2024.1
Requests==2.32.3
waitress==3.0.0
Werkzeug==3.0.4
